requirements.txt
absl-py==1.2.0
aggdraw==1.3.14
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
graphviz==0.20
h5py==3.6.0
importlib-metadata==4.11.3
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
lxml==4.9.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
pyqt5-plugins==5.15.4.2.2
pyqt5-tools==5.15.4.3.2
pytz==2022.1
pywin32==304
pywinpty==2.0.6
qt5-tools==5.15.2.1.2
qtconsole==5.3.1
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sklearn==0.0
tensorflow-cpu==2.8.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9

